Add TimeOverview component to EventBanner
